// src/knowledge/ast/models/NodeKind.ts
export enum NodeKind {
  Repository = 'Repository',
  Module = 'Module',
  Namespace = 'Namespace',
  File = 'File',
  Class = 'Class',
  Interface = 'Interface',
  Enum = 'Enum',
  Function = 'Function',
  Method = 'Method',
  Variable = 'Variable',
  Constant = 'Constant',
  Parameter = 'Parameter',
  Property = 'Property',
  Import = 'Import',
  Export = 'Export',
  Decorator = 'Decorator',
  Annotation = 'Annotation',
  Documentation = 'Documentation',
  Comment = 'Comment',
  Statement = 'Statement',
  Expression = 'Expression',
  Unknown = 'Unknown'
}